+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project employs a versioning scheme described in [RELEASE.md](RELEASE.md#ve
 
 ## [Unreleased]
 
+## [1.23.4] - 2022-10-21
+
 ### Changed
 
 - **\[Action Required\]** Don't use tainted node as control plane node (#572) \
@@ -85,7 +87,8 @@ This project employs a versioning scheme described in [RELEASE.md](RELEASE.md#ve
 - See [release-1.13/CHANGELOG.md](https://github.com/cybozu-go/cke/blob/release-1.13/CHANGELOG.md) for changes in CKE 1.13.
 - See [release-1.12/CHANGELOG.md](https://github.com/cybozu-go/cke/blob/release-1.12/CHANGELOG.md) for changes in CKE 1.12.
 
-[Unreleased]: https://github.com/cybozu-go/cke/compare/v1.23.3...HEAD
+[Unreleased]: https://github.com/cybozu-go/cke/compare/v1.23.4...HEAD
+[1.23.4]: https://github.com/cybozu-go/cke/compare/v1.23.3...v1.23.4
 [1.23.3]: https://github.com/cybozu-go/cke/compare/v1.23.2...v1.23.3
 [1.23.2]: https://github.com/cybozu-go/cke/compare/v1.23.1...v1.23.2
 [1.23.1]: https://github.com/cybozu-go/cke/compare/v1.23.0...v1.23.1